cleaned up a bit
git-svn-id: https://scm.mmbase.org/mmbase/trunk@45316 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ src/main/java/org/mmbase/util/ExprCalc.java
@@ -93,7 +93,7 @@ public class ExprCalc {
                     log.error("Could not evaluate expression '" + token + "' of '" + input + "'");
                 }
                 try {
-                    mNum=(Double.valueOf(token)).doubleValue();
+                    mNum= Double.valueOf(token);
                 } catch (NumberFormatException e) {
                     log.error("Could not evaluate expression ('" + token + "' not a number) of '" + input + "'");
                 }
